Update Swashbuckle.AspNetCore to version 6.8.0
Updated the `MinimalSample.csproj` file to change the version of the `Swashbuckle.AspNetCore` package from `6.7.3` to `6.8.0`.
Updated the `MinimalHelpers.OpenApi.csproj` file to change the version of the `Swashbuckle.AspNetCore.SwaggerGen` package from `6.7.3` to `6.8.0`.

These updates likely include bug fixes, new features, or other improvements provided in the newer version of the packages.